refactor: 피드백 반영 - step3 패키지 삭제, 패키지 이름 변경(step4 -> racingcar)
