+++ src/main/java/me/namuhuchutong/racingcar/domain/rule/RacingRule.java
@@ -1,0 +1,6 @@
+package me.namuhuchutong.racingcar.domain.rule;
+
+@FunctionalInterface
+public interface RacingRule {
+    boolean isSatisfiedRule();
+}
