--- src/main/java/me/namuhuchutong/step3/DefaultRacingRule.java
@@ -1,9 +1,0 @@
-package me.namuhuchutong.step3;
-
-public class DefaultRacingRule implements RacingRule {
-
-    @Override
-    public boolean isSatisfiedRule() {
-        return true;
-    }
-}
