--- src/main/java/me/namuhuchutong/step3/README.md
@@ -1,34 +1,0 @@
-# 3단계 - 자동차 경주
-
-## 기능 요구사항
-
----
-
-### 1. 자동차
-
-- 자동차 수는 사용자가 결정할 수 있다.
-    - 자동차 수가 음수이거나 0이면 프로그램이 종료된다.
-- 자동차의 경주 횟수는 사용자가 결정할 수 있다.
-    - 경주 횟수가 음수이거나 0이면 프로그램이 종료된다.
-- 자동차는 랜덤한 값에 따라 움직일 수 있다.
-    - 값은 0~9 사이에서 랜덤하게 가져온다.
-    - 값이 4미만인 경우 자동차는 멈춰있다.
-    - 값이 4이상일 경우 자동차는 전진한다.
-    - 자동차는 전진할 때 1칸 전진한다.
-    - 자동차의 전진은 "-"으로 표시한다.
-
-### 2. 화면
-
-- 자동차 경주는 콘솔로 출력이 된다.
-    - 숫자 이외의 입력이 있을 경우 프로그램이 종료된다. 
-- 어느 시점에 출력할 것인지에 대한 제약은 없다.
-
-## 프로그래밍 요구사항
-
----
-
-- 핵심 로직과 UI 로직은 구분되어야 한다.
-- 모든 로직에 단위 테스트를 구현한다.
-    - 단, UI 로직은 제외한다.
-- 자바 코드 컨벤션을 지키면서 프로그래밍한다.
-- else를 사용하지 않는다.
