--- src/main/java/me/namuhuchutong/step3/RacingRule.java
@@ -1,7 +1,0 @@
-package me.namuhuchutong.step3;
-
-@FunctionalInterface
-public interface RacingRule {
-
-    boolean isSatisfiedRule();
-}
