--- src/main/java/me/namuhuchutong/step4/domain/rule/RacingRule.java
@@ -1,6 +1,0 @@
-package me.namuhuchutong.step4.domain.rule;
-
-@FunctionalInterface
-public interface RacingRule {
-    boolean isSatisfiedRule();
-}
